--- voteSite/board/views.py
@@ -1,20 +1,0 @@
-from .models import Board
-from .serializers import BoardSerializer
-from rest_framework import generics, permissions
-from .permission import IsOwnerOrReadOnly
-
-
-class BoardList(generics.ListCreateAPIView):
-    queryset = Board.objects.all()
-    serializer_class = BoardSerializer
-    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
-
-    def perform_create(self, serializer):
-        serializer.save(owner=self.request.user)
-
-
-class BoardDetail(generics.RetrieveUpdateDestroyAPIView):
-    queryset = Board.objects.all()
-    serializer_class = BoardSerializer
-    permission_classes = [permissions.IsAuthenticatedOrReadOnly, IsOwnerOrReadOnly]
-
