+++ voteSite/config/urls.py
@@ -9,7 +9,7 @@ urlpatterns = [
     # path('vote/', include('vote.urls')),
     path('', include('users.urls')),
     path('comments/', include('comments.urls')),
-    path('boards/', include('board.urls'))
+    path('boards/', include('boards.urls'))
 ]
 urlpatterns += [
     path('api-auth/', include('rest_framework.urls')),
